+++ jakarta.batch.arquillian.extension/src/main/java/ee/jakarta/tck/batch/arquillian/ArquillianExtension.java
@@ -14,7 +14,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jbatch.arquillian.extension;
+package ee.jakarta.tck.batch.arquillian;
 
 import org.jboss.arquillian.container.test.spi.client.deployment.DeploymentScenarioGenerator;
 import org.jboss.arquillian.core.spi.LoadableExtension;
