+++ jakarta.batch.arquillian.extension/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-jbatch.arquillian.extension.ArquillianExtension
+ee.jakarta.tck.batch.arquillian.ArquillianExtension
