+++ jakarta.batch.reporting/src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,1 @@
-jakarta.batch.junit5reporting.MyTestExecutionListener
+ee.jakarta.tck.batch.junit5.reporting.MyTestExecutionListener
